+++ views/single-user.handlebars
@@ -1,27 +1,72 @@
 <article class="row d-block w-100 mt-5">
 <form>
+  <h3>Personal Information</h3>
   <div class="form-row">
     <div class="form-group col-md-6">
-      <label for="inputFirstName">First Name</label>
+      <label for="userFirstName">First Name</label>
       <input type="text" class="form-control" id="userFirstName" value="{{ user.first_name }}" />
     </div>
     <div class="form-group col-md-6">
-      <label for="inputLastName">Last Name</label>
+      <label for="userLastName">Last Name</label>
       <input type="text" class="form-control" id="userLastName" value="{{ user.last_name }}" />
     </div>
   </div>
-    <div class="form-row">
+  <div class="form-row">
     <div class="form-group col-md-6">
-      <label for="inputUserUsername">Username</label>
-      <input type="text" class="form-control" id="inputUserUsername" value="{{ user.username }}" />
+      <label for="userDob">Date of Birth</label>
+      <input type="date" class="form-control" id="userDob" value="{{ user.dob }}" />
     </div>
     <div class="form-group col-md-6">
-      <label for="inputUserPassword">Password</label>
-      <input type="password" class="form-control" id="inputUserPassword" value="{{ user.password }}" />
+      <label for="userSsn">Social Security #</label>
+      <input type="text" class="form-control" id="userSsn" value="{{ user.ssn }}" />
     </div>
   </div>
+  <div class="form-row">
+    <div class="form-group col-md-6">
+      <label for="userUsername">Username</label>
+      <input type="text" class="form-control" id="userUsername" value="{{ user.username }}" />
+    </div>
+    <div class="form-group col-md-6">
+      <label for="userPassword">Password</label>
+      <input type="password" class="form-control" id="userPassword" value="{{ user.password }}" />
+    </div>
+  </div>
+  <div class="form-row">
+    <div class="form-group col-md-2">
+      <div class="form-check">
+        <input class="form-check-input" type="checkbox" id="userActive" value={{ user.active }}>
+        <label class="form-check-label" for="userActive">
+          Active
+        </label>
+      </div>
+    </div>
+    <div class="form-group col-md-4">
+      <label for="userAccessType">Access Type</label>
+      <input type="text" class="form-control" id="userAccessType" value="{{ user.access.access_type }}" />
+    </div>
+    <div class="form-group col-md-6">
+      <label for="inputUserAccessDesc">Description</label>
+      <input type="text" class="form-control" id="inputUserAccessDesc" value="{{ user.access.access_desc }}" />
+    </div>
+  </div>
+  <h3>Contact Information</h3>
   <div class="form-group">
-    <label for="inputUserAddress">Address</label>
+    <label for="inputUserEmail">Email</label>
+    <input type="email" class="form-control" id="inputUserEmail" value="{{ user.email }}">
+  </div>
+  <div class="form-row">
+    <div class="form-group col-md-6">
+      <label for="inputUserPrimaryPhone">Mobile Phone</label>
+      <input type="phone" class="form-control" id="inputUserPrimaryPhone" value="{{ user.primary_phone }}" />
+    </div>
+    <div class="form-group col-md-6">
+      <label for="inputUserAltPhone">Other Phone</label>
+      <input type="phone" class="form-control" id="inputUserPassword" value="{{ user.alt_phone }}" />
+    </div>
+  </div>
+  <h4>Address:</h4>
+  <div class="form-group">
+    <label for="inputUserAddress">Street Address</label>
     <input type="text" class="form-control" id="inputUserAddress" value="{{ user.street_address }}">
   </div>
   <div class="form-row">
@@ -41,14 +86,7 @@
       <input type="text" class="form-control" id="inputUserZip" value="{{ user.zip }}">
     </div>
   </div>
-  <div class="form-group">
-    <div class="form-check">
-      <input class="form-check-input" type="checkbox" id="userActive" value={{ user.active }}>
-      <label class="form-check-label" for="userActive">
-        Active
-      </label>
-    </div>
-  </div>
+  <h3>Clinician Information</h3>
   <button type="submit" class="btn btn-primary">Save</button>
 </form>
 </article>
